Refactor CMake build system and host files
Use CMakeLists.txt file for each directory. Setup to be consumed by
monad-linux's CMake. Utilize GoogleTest's CMake. Refactor host files
layout.
